유저 게시글만 검색되게 수정, 로딩화면, 비밀번호 해싱, 글수정후 뷰페이지 이동 수정

+++ src/components/Cancel.tsx
@@ -1,36 +1,38 @@
-'use client';
+"use client";
 
-import React from 'react';
+import React from "react";
 import { useRouter } from "next/navigation";
 
 interface CancelProps {
-    id: string;
+id: string;
 }
 
 const Cancel: React.FC<CancelProps> = ({ id }) => {
-    const router = useRouter();
+const router = useRouter();
 
-    const handleCancel = async () => {
-        try {
-            const res = await fetch("/api/auth/delete", {
-                method: "DELETE",
-                headers: {
-                    "Content-Type": "application/json",
-                },
-                body: JSON.stringify({ _id: id }),
-            });
-            console.log(res);
-            window.location.reload();
-        } catch (error) {
-            console.error('회원탈퇴 실패', error);
-        }
-    };
+const handleCancel = async () => {
+try {
+    if (confirm("회원 탈퇴 하시겠습니까?")) {
+    const res = await fetch("/api/auth/delete", {
+        method: "DELETE",
+        headers: {
+        "Content-Type": "application/json",
+        },
+        body: JSON.stringify({ _id: id }),
+    });
+    console.log(res);
+    window.location.reload();
+    }
+} catch (error) {
+    console.error("회원탈퇴 실패", error);
+}
+};
 
-    return (
-        <button onClick={handleCancel} className='text-sm'>
-            회원탈퇴
-        </button>
-    );
+return (
+<button onClick={handleCancel} className="text-sm">
+    회원탈퇴
+</button>
+);
 };
 
 export default Cancel;
